msm: camerav2: sensor: Avoid csiphy release when used in combo mode
When two camera sensors are running concurrently and sharing csiphy
core dont release csiphy core when one of the sensor is streamoff.
This will cut off the mipi data for other sensor which is already
streaming on same core. Release csiphy core only one sensor is in
use for that instant.

Change-Id: I3fde0d122ce46ac6f95d7d0c1cc65ae9377d8f3d
Signed-off-by: Lokesh Kumar Aakulu <lkumar@codeaurora.org>
Signed-off-by: Vivek Veenam <vveenam@codeaurora.org>