ARM: dts: msm: Add support for SDIO in sdhc2 for apq8017 config-1/3/8
Add entries for support of sdio in apq8017. The properties
different from sd card are deleted or overwritten and properties
required for SDIO are added to the relevant files of 8017 config 1,3
and 8. The QCA and SDIO IO voltage regulators are automatically
turned on all the time. So SDIO doesn't vote for any VDD but
still votes for the regulator of SDIO IO.

Change-Id: Ie5a49f28b01972881dffe501c8578eb983e7ef98
Signed-off-by: Vijay Viswanath <vviswana@codeaurora.org>