mmc: mmc: Handle error case in mmc_suspend
Currently in case if suspend is failed due to some
reason then the error case is not handled properly :-
1. w.r.t. suspend_clk_scaling claim_host counters are going bad.
2. In case of any error after halting and disabling cmdq -
the error is returned to the PM framework and it assumes that
card is not suspended, whereas CQ remains disabled. This causes
I/O requests hangs since no request can be processed any further.

Fix this by handling error cases properly.

Change-Id: I8691eebcb0e2d089720505475aa0297efce5cca5
Signed-off-by: Ritesh Harjani <riteshh@codeaurora.org>
Signed-off-by: Sayali Lokhande <sayalil@codeaurora.org>