msm: kgsl: Relocate the pagefault tracepoint
During a pagefault we print trace with the fault
information which gets delayed by unacceptable
time. It happens due to the the code that searches the
faulty address in the recently freed addresses for a matching
entry.

Moving it to new position wont have any impact on debugging
instead trace is dumped without any delay.

Change-Id: I11487798b60742096e06605cf3046b4cfc8cc057
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>