usb: gadget: f_mbim: Fix incorrect repsonse available notifications
Two notify requests are being queued for one available modem
response. So for the first GET_ENCAPSULATED_RESPONSE we provide
the host with the available response. Now for the next
GET_ENCAPSULATED_COMMAND we notify the host that the response is
available even before the modem is ready with a response because
of the extra notify request queued on the interrupt endpoint.
This causes a STALL for the next GET_ENCAPSULATED_RESPONSE request.
This is caused because we are queueing a notify request from the
completion handler of the interrupt endpoint request when the
response queue is not empty.

Change-Id: I8a018dc5a95695322fbae096da5943fd3f648f1f
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>