msm: ipa: Fix to incorrect dl stats
DL stats from Q6 do not include the additional LAN/WLAN
header bytes for respective endpoint. Make a change to
not to deduct the header bytes to keep the stats
accurate.

Also make a change to not to send QMI message to Q6
when there are no clients connected to avoid unnecessary
QMI exchange with Q6.

Change-Id: I3e7b34b106053c6d902ac0aafa9a362c1de0e221
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Utkarsh Saxena <usaxena@codeaurora.org>