clk: msm: Change the halt check type for gcc_pcie_pipe_clk for mdm9x40
The gcc_pcie_pipe_clk clock has an external source. Due to HW design,
there is no guarantee that 1) the downstream branch clocks are turned on
after turning on the sources; 2) the downstream branch clocks are turned
off before turning off the sources.
Therefore, make the halt check for this clock as delay instead of
polling the status bit to avoid warnings.

Change-Id: I0913e94a642dcec5d049b79524401f39956b8a88
Signed-off-by: Amit Nischal <anischal@codeaurora.org>